Restore master to original package
