+++ business_rules/utils.py
@@ -1,10 +1,8 @@
-import inspect
 from decimal import Decimal, Inexact, Context
-
+import inspect
 
 def fn_name_to_pretty_label(name):
     return ' '.join([w.title() for w in name.split('_')])
-
 
 def export_rule_data(variables, actions):
     """ export_rule_data is used to export all information about the
@@ -19,13 +17,12 @@ def export_rule_data(variables, actions):
     variables_data = variables.get_all_variables()
     variable_type_operators = {}
     for variable_class in inspect.getmembers(operators, lambda x: getattr(x, 'export_in_rule_data', False)):
-        variable_type = variable_class[1]  # getmembers returns (name, value)
+        variable_type = variable_class[1] # getmembers returns (name, value)
         variable_type_operators[variable_type.name] = variable_type.get_all_operators()
 
     return {"variables": variables_data,
             "actions": actions_data,
             "variable_type_operators": variable_type_operators}
-
 
 def float_to_decimal(f):
     """
@@ -43,12 +40,3 @@ def float_to_decimal(f):
         result = ctx.divide(numerator, denominator)
     return result
 
-
-def get_difference(all_dicts, dict_to_subtract):
-    """
-
-    :param all_dicts:
-    :param dict_to_subtract:
-    :return:
-    """
-    return filter(lambda single_dict: single_dict not in dict_to_subtract, all_dicts)
